requirements.txt
# FastAPI + Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis + Celery
redis>=5.0.0
celery[redis]>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# HTTP + Scraping
aiohttp>=3.9.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# PDF + OCR
pdfplumber>=0.11.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.2.0

# NLP
spacy>=3.7.0

# Data
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.28.0

# Dev
black>=24.1.0
ruff>=0.2.0
